src/api/requirements.txt
flask
sqlalchemy
mysqlclient
tensorflow
pyjwt